replaced default implementation, introduced legacy version
